woow/services/ui_gateway/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
requests==2.32.3
